Update example game with main cs files
